# public/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "NEO APP API",
    "version": "1.0.0",
    "description": "REST API ของร้าน NEO APP (ขายแอปพลิเคชัน + กระเป๋าเงิน)\n\n## การยืนยันตัวตน\n\nทุก endpoint ต้องผ่านการยืนยันตัวตน ไม่มีเส้นไหนเปิดสาธารณะ รองรับ 2 วิธี:\n\n1. **Cookie session (เบราว์เซอร์)** — ล็อกอินที่ `/login` แล้ว Supabase จะเก็บ JWT ไว้ใน cookie `sb-<project-ref>-auth-token`\n2. **Bearer token (Postman / curl / script)** — ส่งเฮดเดอร์ `Authorization: Bearer <access_token>`\n\nขอ access token:\n\n```bash\ncurl -X POST \"$NEXT_PUBLIC_SUPABASE_URL/auth/v1/token?grant_type=password\" \\\n  -H \"apikey: $NEXT_PUBLIC_SUPABASE_PUBLISHABLE_KEY\" \\\n  -H \"content-type: application/json\" \\\n  -d '{\"email\":\"you@example.com\",\"password\":\"...\"}'\n```\n\n## สิทธิ์\n\nข้อมูลถูกคุมด้วย Row Level Security ของ Postgres: ลูกค้าเห็นเฉพาะคำสั่งซื้อ/กระเป๋าเงินของตัวเอง\nส่วนการเพิ่ม-แก้-ลบสินค้า อัปโหลดรูป และแก้ตั้งค่าร้าน ทำได้เฉพาะบัญชีที่มี `app_metadata.role = \"admin\"`\n\n## เอกสารนี้\n\nหน้า `/docs` และไฟล์ `/openapi.json` เปิดให้เฉพาะแอดมิน คนอื่นจะได้ 404",
    "contact": {
      "name": "NEO APP"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local development"
    }
  ],
...
  "security": [
    {
      "cookieAuth": []
    },
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/api/products/{id}": {
      "parameters": [
        {
          "name": "id",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string"
          },
          "example": "prod-1"
        }
      ],
      "get": {
        "tags": [
          "Products"
        ],
        "summary": "ข้อมูลสินค้ารายชิ้น",
        "operationId": "getProduct",
        "responses": {
          "200": {
            "description": "สำเร็จ",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        }
      },
      "patch": {
        "tags": [
          "Products"
        ],
        "summary": "แก้ไขข้อมูลสินค้า (ส่งเฉพาะฟิลด์ที่ต้องการเปลี่ยน)",
        "operationId": "updateProduct",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProductInput"
              },
              "example": {
                "price": 2490,
                "stock": 12
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "อัปเดตสำเร็จ",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        }
      },
      "delete": {
        "tags": [
          "Products"
        ],
        "summary": "ลบสินค้า",
        "operationId": "deleteProduct",
        "responses": {
          "200": {
            "description": "ลบสำเร็จ",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        }
      }
    },
    "/api/products/{id}/reviews": {
      "parameters": [
        {
          "name": "id",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string"
          },
          "example": "prod-1"
        }
      ],
      "post": {
        "tags": [
          "Products"
        ],
        "summary": "เพิ่มรีวิวสินค้า",
        "description": "ชื่อผู้รีวิวมาจากบัญชีที่ล็อกอิน ไม่ได้รับค่าจาก request body",
        "operationId": "createProductReview",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "rating",
                  "comment"
                ],
                "properties": {
                  "rating": {
                    "type": "number",
                    "minimum": 1,
                    "maximum": 5,
                    "example": 5
                  },
                  "comment": {
                    "type": "string",
                    "example": "เสียงดีมาก คุ้มราคา"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "ส่งรีวิวสำเร็จ",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "$ref": "#/components/schemas/Review"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        }
      }
    },
    "/api/orders": {
      "get": {
        "tags": [
          "Orders"
        ],
        "summary": "รายการคำสั่งซื้อ",
        "operationId": "listOrders",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "description": "กรองตามสถานะ ใช้ `ทั้งหมด` เพื่อไม่กรอง",
            "schema": {
              "$ref": "#/components/schemas/OrderStatus"
            }
          },
          {
            "name": "search",
            "in": "query",
            "description": "ค้นหาจากเลขที่คำสั่งซื้อ ชื่อลูกค้า หรือเบอร์โทร",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "สำเร็จ",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "count": {
                      "type": "integer"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Order"
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        }
      },
      "post": {
        "tags": [
          "Orders"
        ],
        "summary": "สั่งซื้อและตัดเงินจากกระเป๋า",
        "operationId": "createOrder",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "customer",
                  "items"
                ],
                "properties": {
                  "customer": {
                    "$ref": "#/components/schemas/CustomerInfo"
                  },
                  "items": {
                    "type": "array",
                    "minItems": 1,
                    "items": {
                      "type": "object",
                      "required": [
                        "productId",
                        "quantity"
                      ],
                      "properties": {
                        "productId": {
                          "type": "string",
                          "format": "uuid"
                        },
                        "quantity": {
                          "type": "integer",
                          "minimum": 1
                        },
                        "selectedColor": {
                          "type": "string"
                        }
                      }
                    }
                  },
                  "couponCode": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "สร้างคำสั่งซื้อสำเร็จ",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        },
        "description": "เซิร์ฟเวอร์คิดราคาจากฐานข้อมูลเอง ตรวจสต็อก ตัดเงินในกระเป๋า และสร้างคำสั่งซื้อในทรานแซกชันเดียว ราคาที่ส่งมาจากเบราว์เซอร์ไม่ถูกใช้\n\nถ้าสินค้ามาจากซัพพลายเออร์ ระบบจะสั่งซื้อต่อให้อัตโนมัติและคืน account มาพร้อม response หากสั่งต่อไม่สำเร็จจะคืนเงินเข้ากระเป๋าให้ทันที"
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "sb-access-token",
        "description": "Cookie session ของ Supabase (ชื่อจริงคือ `sb-<project-ref>-auth-token`) เบราว์เซอร์ส่งให้อัตโนมัติหลังล็อกอิน"
      },
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Supabase access token (JWT) สำหรับ API client ที่ไม่มี cookie jar"
      }
    },
    "responses": {
      "Unauthorized": {
        "description": "ยังไม่ได้ล็อกอิน หรือ token หมดอายุ/ไม่ถูกต้อง",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "success": false,
              "error": "unauthorized",
              "message": "กรุณาเข้าสู่ระบบก่อนเรียกใช้งาน API"
            }
          }
        }
      },
      "BadRequest": {
        "description": "ข้อมูลที่ส่งมาไม่ครบหรือไม่ถูกต้อง",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
      "NotFound": {
        "description": "ไม่พบข้อมูลตาม ID ที่ระบุ",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
      "ServerError": {
        "description": "ข้อผิดพลาดฝั่งเซิร์ฟเวอร์",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      }
    },
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": false
          },
          "error": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        }
      },
      "Review": {
        "type": "object",
        "required": [
          "id",
          "userName",
          "rating",
          "comment",
          "date"
        ],
        "properties": {
          "id": {
            "type": "string",
            "example": "rev-1"
          },
          "userName": {
            "type": "string",
            "example": "คุณสมชาย ใจดี"
          },
          "rating": {
            "type": "number",
            "minimum": 1,
            "maximum": 5,
            "example": 5
          },
          "comment": {
            "type": "string"
          },
          "date": {
            "type": "string",
            "example": "2026-08-14"
          }
        }
      },
      "Product": {
        "type": "object",
        "required": [
          "id",
          "name",
          "category",
          "price",
          "stock",
          "description",
          "specs",
          "image",
          "rating",
          "reviewsCount"
        ],
        "properties": {
          "id": {
            "type": "string",
            "example": "3f6d1c2e-0f1a-4d3b-9a7c-2b8e5d4f1a90"
          },
          "name": {
            "type": "string"
          },
          "category": {
            "type": "string"
          },
          "price": {
            "type": "number"
          },
          "originalPrice": {
            "type": "number"
          },
          "stock": {
            "type": "integer"
          },
          "description": {
            "type": "string"
          },
          "specs": {
            "type": "object",
            "additionalProperties": {
              "type": "string"
            }
          },
          "image": {
            "type": "string",
            "format": "uri"
          },
          "gallery": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uri"
            }
          },
          "rating": {
            "type": "number"
          },
          "reviewsCount": {
            "type": "integer"
          },
          "badge": {
            "type": "string",
            "enum": [
              "HOT",
              "NEW",
              "SALE",
              "LIMITED"
            ]
          },
          "reviews": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Review"
            }
          },
          "isFeatured": {
            "type": "boolean"
          }
        }
      },
      "ProductInput": {
        "type": "object",
        "required": [
          "name",
          "price",
          "stock"
        ],
        "properties": {
          "name": {
            "type": "string"
          },
          "category": {
            "type": "string"
          },
          "price": {
            "type": "number"
          },
          "originalPrice": {
            "type": "number"
          },
          "stock": {
            "type": "integer"
          },
          "description": {
            "type": "string"
          },
          "specs": {
            "type": "object",
            "additionalProperties": {
              "type": "string"
            }
          },
          "image": {
            "type": "string",
            "format": "uri"
          },
          "badge": {
            "type": "string",
            "enum": [
              "HOT",
              "NEW",
              "SALE",
              "LIMITED"
            ]
          },
          "isFeatured": {
            "type": "boolean"
          }
        }
      },
      "ProductResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "$ref": "#/components/schemas/Product"
          }
        }
      },
      "CustomerInfo": {
        "type": "object",
        "required": [
          "name",
          "email",
          "phone",
          "address",
          "district",
          "province",
          "postalCode"
        ],
        "properties": {
          "name": {
            "type": "string",
            "example": "คุณสมชาย ใจดี"
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "phone": {
            "type": "string",
            "example": "081-987-6543"
          },
          "address": {
            "type": "string"
          },
          "district": {
            "type": "string"
          },
          "province": {
            "type": "string"
          },
          "postalCode": {
            "type": "string",
            "example": "10110"
          },
          "note": {
            "type": "string"
          }
        }
      },
      "OrderStatus": {
        "type": "string",
        "enum": [
          "รอดำเนินการ",
          "กำลังจัดเตรียม",
          "จัดส่งแล้ว",
          "สำเร็จ",
          "ยกเลิก"
        ]
      },
      "PaymentMethod": {
        "type": "string",
        "enum": [
          "wallet",
          "promptpay",
          "credit_card",
          "bank_transfer",
          "cod"
        ]
      },
      "Order": {
        "type": "object",
        "required": [
          "id",
          "createdAt",
          "customer",
          "items",
          "subtotal",
          "discount",
          "shippingFee",
          "totalAmount",
          "status",
          "paymentMethod",
          "isPaid"
        ],
        "properties": {
          "id": {
            "type": "string",
            "example": "ORD-260814-A7F3C"
          },
          "createdAt": {
            "type": "string",
            "example": "2026-08-14 09:30"
          },
          "customer": {
            "$ref": "#/components/schemas/CustomerInfo"
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OrderItem"
            }
          },
          "subtotal": {
            "type": "number"
          },
          "discount": {
            "type": "number"
          },
          "shippingFee": {
            "type": "number"
          },
          "totalAmount": {
            "type": "number"
          },
          "status": {
            "$ref": "#/components/schemas/OrderStatus"
          },
          "paymentMethod": {
            "$ref": "#/components/schemas/PaymentMethod"
          },
          "isPaid": {
            "type": "boolean"
          },
          "trackingNumber": {
            "type": "string"
          },
          "couponCode": {
            "type": "string"
          }
        }
      },
      "OrderResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "$ref": "#/components/schemas/Order"
          }
        }
      },
...
      "OrderItem": {
        "type": "object",
        "description": "สำเนาข้อมูลสินค้า ณ เวลาที่สั่งซื้อ แก้สินค้าภายหลังไม่กระทบคำสั่งซื้อเดิม",
        "required": [
          "productId",
          "name",
          "unitPrice",
          "quantity"
        ],
        "properties": {
          "productId": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "image": {
            "type": "string",
            "format": "uri"
          },
          "unitPrice": {
            "type": "number"
          },
          "quantity": {
            "type": "integer"
          },
          "selectedColor": {
            "type": "string"
          }
        }
      },
...
    }
  }
}